+++ packages/server/test/react.test.tsx
@@ -34,10 +34,10 @@ import {
 import { withTRPC } from '../../next/src';
 import { OutputWithCursor, createReactQueryHooks } from '../../react/src';
 import { createSSGHelpers } from '../../react/src/ssg';
-import { DefaultErrorShape } from '../src';
 import { TRPCError } from '../src/TRPCError';
-import { getErrorFromUnknown } from '../src/internals/errors';
-import { Observable, Observer, observable } from '../src/observable';
+import { observable } from '../src/observable';
+import { DefaultErrorShape } from '../src/router';
+import { subscriptionPullFactory } from '../src/subscription';
 
 setLogger({
   log() {},
@@ -51,42 +51,6 @@ type Post = {
   title: string;
   createdAt: number;
 };
-
-function subscriptionPullFactory<TOutput>(opts: {
-  /**
-   * The interval of how often the function should run
-   */
-  intervalMs: number;
-  pull(emit: Observer<TOutput, unknown>): void | Promise<void>;
-}): Observable<TOutput, unknown> {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let timer: any;
-  let stopped = false;
-  async function _pull(emit: Observer<TOutput, unknown>) {
-    /* istanbul ignore next */
-    if (stopped) {
-      return;
-    }
-    try {
-      await opts.pull(emit);
-    } catch (err /* istanbul ignore next */) {
-      emit.error(getErrorFromUnknown(err));
-    }
-
-    /* istanbul ignore else */
-    if (!stopped) {
-      timer = setTimeout(() => _pull(emit), opts.intervalMs);
-    }
-  }
-
-  return observable<TOutput>((emit) => {
-    _pull(emit).catch((err) => emit.error(getErrorFromUnknown(err)));
-    return () => {
-      clearTimeout(timer);
-      stopped = true;
-    };
-  });
-}
 
 function createAppRouter() {
   const db: {
